requirements.txt
requests>=2.31.0
pyyaml>=6.0
aiohttp>=3.9.0
asyncio>=3.4.3
base64
urllib3>=2.0.0
cloudscraper>=1.2.71

